actualización

+++ src/main/java/org/springframework/samples/dobble/card/CardRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CardRepository extends CrudRepository<Card, Integer> { 
+public interface CardRepository extends CrudRepository<Card, Long> { 
     
 }
